Renamed the typed vectors from FooVector2 to just FooVector. This is a 2d, game, so I do not care. In the distant future, I suppose I might make it more general, but seriously, fuck enhancing this any further.
oh, and they use a vector internally instead of 2 floats. Dunno which way is better admittedly, but since I am manipulating it as a vector wrapper, storing a vector is probably better.

And, a global "Rename" alias so you can specify vector or point, depending on if it is a movement thinggy or a position thinggy. Same object either way, just syntax helper.